Revert "Add the support of certificate checking for secure mode (#39)"
This reverts commit cef179c81f07f86413c720623995c17bd1bddd7d.
