+++ whisk/client_test.go
@@ -25,7 +25,6 @@ import (
     "net/http"
     "fmt"
     "net/url"
-    "crypto/tls"
 )
 
 const (
@@ -37,20 +36,18 @@ const (
     FakeAuthKey = "dhajfhshfs:hjhfsjfdjfjsgfjs"
 )
 
-func GetValidConfigTest(insecure bool) *Config {
+func GetValidConfigTest() *Config {
     var config Config
     config.Host = FakeHost
     config.Namespace = FakeNamespace
     config.AuthToken = FakeAuthKey
-    config.Insecure = insecure
     return &config
 }
 
-func GetInvalidConfigMissingApiHostTest(insecure bool) *Config {
+func GetInvalidConfigMissingApiHostTest() *Config {
     var config Config
     config.Namespace = FakeNamespace
     config.AuthToken = FakeAuthKey
-    config.Insecure = insecure
     return &config
 }
 
@@ -63,20 +60,19 @@ func GetInvalidConfigMissingApiHostWithBaseURLTest() *Config {
     return &config
 }
 
-func GetValidConfigDiffApiHostAndBaseURLTest(insecure bool) *Config {
+func GetValidConfigDiffApiHostAndBaseURLTest() *Config {
     var config Config
     urlBase := fmt.Sprintf("https://%s/api", FakeHostDiff)
     config.BaseURL, _ = url.Parse(urlBase)
     config.Host = FakeHost
     config.Namespace = FakeNamespace
     config.AuthToken = FakeAuthKey
-    config.Insecure = insecure
     return &config
 }
 
-func TestNewClientDisablingCertificate(t *testing.T) {
+func TestNewClient(t *testing.T) {
     // Test the use case to pass a valid config.
-    config := GetValidConfigTest(true)
+    config := GetValidConfigTest()
     client, err := NewClient(http.DefaultClient, config)
     assert.Nil(t, err)
     assert.NotNil(t, client)
@@ -86,7 +82,7 @@ func TestNewClientDisablingCertificate(t *testing.T) {
     assert.Equal(t, FakeAuthKey, client.Config.AuthToken)
 
     // Test the use case to pass an invalid config with a missing api host.
-    config = GetInvalidConfigMissingApiHostTest(true)
+    config = GetInvalidConfigMissingApiHostTest()
     client, err = NewClient(http.DefaultClient, config)
     assert.NotNil(t, err)
     assert.Contains(t, err.Error(), "Unable to create request URL, because OpenWhisk API host is missing")
@@ -100,7 +96,7 @@ func TestNewClientDisablingCertificate(t *testing.T) {
     assert.Nil(t, client)
 
     // Test the use case to pass a valid config with both the base and api host of different values.
-    config = GetValidConfigDiffApiHostAndBaseURLTest(true)
+    config = GetValidConfigDiffApiHostAndBaseURLTest()
     client, err = NewClient(http.DefaultClient, config)
     assert.Nil(t, err)
     assert.NotNil(t, client)
@@ -108,45 +104,4 @@ func TestNewClientDisablingCertificate(t *testing.T) {
     assert.Equal(t, FakeHost, client.Config.Host)
     assert.Equal(t, FakeBaseURLDiff, client.Config.BaseURL.String())
     assert.Equal(t, FakeAuthKey, client.Config.AuthToken)
-}
-
-func TestNewClientEnablingCertificate(t *testing.T) {
-    TEST_KEY_FILE := "TEST_KEY_FILE"
-    TEST_CERT_FILE := "TEST_CERT_FILE"
-
-    // Test the use case to pass a config in secure mode missing the cert and key files.
-    config := GetValidConfigTest(false)
-    _, err := NewClient(http.DefaultClient, config)
-    assert.NotNil(t, err)
-
-    // Test the use case to pass a config in secure mode with non-existing the cert and key files.
-    config = GetValidConfigTest(false)
-    config.Key = TEST_KEY_FILE
-    config.Cert = TEST_CERT_FILE
-    _, err = NewClient(http.DefaultClient, config)
-    assert.NotNil(t, err)
-
-    // Test the use case to pass a config in secure mode with invalid the cert and key files.
-    CreateFile([]string{ "testKey" }, TEST_KEY_FILE)
-    CreateFile([]string{ "testCert" }, TEST_CERT_FILE)
-    config = GetValidConfigTest(false)
-    config.Key = TEST_KEY_FILE
-    config.Cert = TEST_CERT_FILE
-    _, err = NewClient(http.DefaultClient, config)
-    assert.NotNil(t, err)
-
-    // Test the use case to pass a config in secure mode with valid the cert and key files.
-    oldReadX509KeyPair := ReadX509KeyPair
-    defer func () { ReadX509KeyPair = oldReadX509KeyPair }()
-    ReadX509KeyPair = func(certFile, keyFile string) (tls.Certificate, error) {
-        cert := tls.Certificate{}
-        return cert, nil
-    }
-    config = GetValidConfigTest(false)
-    config.Key = TEST_KEY_FILE
-    config.Cert = TEST_CERT_FILE
-    _, err = NewClient(http.DefaultClient, config)
-    assert.Nil(t, err)
-    DeleteFile(TEST_KEY_FILE)
-    DeleteFile(TEST_CERT_FILE)
 }
