+++ wski18n/i18n_resources.go
@@ -109,12 +109,12 @@ func wski18nResourcesDe_deAllJson() (*asset, error) {
         return nil, err
     }
 
-    info := bindataFileInfo{name: "wski18n/resources/de_DE.all.json", size: 0, mode: os.FileMode(420), modTime: time.Unix(1500156160, 0)}
+    info := bindataFileInfo{name: "wski18n/resources/de_DE.all.json", size: 0, mode: os.FileMode(420), modTime: time.Unix(1506031186, 0)}
     a := &asset{bytes: bytes, info: info}
     return a, nil
 }
 
-var _wski18nResourcesEn_usAllJson = []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\xcc\x58\x5f\x6f\xdb\x36\x10\x7f\xf7\xa7\x38\xf8\xc5\x19\xe0\xea\x03\x74\x4f\xc1\x66\xcc\xc1\xba\xc6\x58\x9d\xf5\x61\x19\x06\x5a\x3c\x47\x87\xc8\x3c\x95\xa4\x9c\xb9\x86\xbe\xfb\x70\x94\x5c\x67\x8d\x65\xfd\xb1\xda\xe5\xc9\x02\xcd\xfb\xdd\x8f\x77\xc7\xfb\xc3\x3f\x47\x00\xfb\x11\x00\xc0\x98\xf4\xf8\x2d\x8c\xef\x8c\x5a\xa5\x08\x9e\x41\x69\x0d\x96\x73\x8f\xc0\x99\x27\x36\x0e\x26\xfb\x7d\x54\x7d\x17\xc5\x64\x3c\x2d\xe5\xbc\x55\xc6\xa5\x4a\x96\x1b\x00\xde\xc2\x73\x80\xf1\x08\xa0\x98\xd6\xeb\x8f\x2d\x2a\x8f\x30\x5f\x2e\x17\x60\xf1\x53\x8e\xce\xc3\x9a\x2d\x2c\xee\x96\x81\x49\x80\x2e\x8a\x49\x40\x45\x6b\x8b\xa2\x91\x51\x0f\xc8\x9e\x24\x7f\x99\x0d\x4e\xf2\x0c\x64\x4f\x92\x3f\xcf\xde\xcd\x96\xb3\xa1\x79\x9e\x47\xed\xeb\xf4\xdb\x0f\xc3\x7b\xfd\x0c\x66\x03\x4d\x95\x65\x68\x74\xcd\xc5\x90\x0d\x77\xbf\xbf\xab\x62\xbf\x27\xe9\xcb\x35\xb4\xb3\xf4\xc1\x20\x02\x27\x40\xb9\x4d\x7b\x59\xb7\x11\xe7\x24\x9d\x1b\xb3\x55\x29\xe9\xbe\x2c\x5a\x8b\x9f\x54\x3e\xb3\x96\x2d\xa0\x89\x59\x93\x79\xf8\x02\xb2\x62\xbd\x6b\xd4\xdc\x4e\xf6\x8c\x5a\x32\xe4\x49\xa5\xf4\xf9\x99\x78\x4b\xad\x0d\xa2\x4d\xa1\xab\x35\xf8\xa4\xba\x0e\x2a\xf7\x09\x1a\x4f\x71\x50\x01\x09\x2a\x8d\xb6\x4b\x94\x76\x01\x3b\x49\xec\x3a\xf7\x09\x5b\xfa\x5c\xca\x3c\xe2\x0e\xc8\x81\x61\x0f\x31\x9b\x35\x3d\xe4\x16\x35\x5c\xbd\x79\x23\xd8\xf2\x8f\x1c\x97\x2c\xea\x1f\x6a\xa8\xf5\x86\x3b\x4d\xce\xc0\xf5\xe2\x06\x12\x76\x1e\x36\xb9\xf8\x17\x21\xb3\xbc\x25\x8d\x3a\xba\x37\x75\x1c\x1a\xa4\x5a\x38\xe8\xfb\xd7\xdd\x9f\x78\xb3\x51\x46\xc3\x5a\x51\x8a\x1a\x74\x5e\x42\x99\xd2\xb5\xb2\x9a\x5b\xac\x51\xdd\x4e\xf6\xa4\xda\xf7\x0c\x64\x3c\xda\xb5\x8a\x8f\x46\xfa\x11\x0c\x1f\xd2\xb5\xcb\xd8\x38\x0c\x17\x0b\xf0\x9f\x0c\x63\x8f\xba\x86\x46\x3f\xac\x6e\xd6\x08\x0a\x8c\x4a\xfb\x5a\xe4\x85\xfc\x49\xf5\xcb\x04\x61\xcd\x69\xca\x4f\x72\xc5\x55\x96\xa5\x87\x4b\x85\x21\x03\x3c\x29\x09\xdd\x18\x69\x8b\xba\xf1\xb6\xf6\x04\x7b\x7d\xf9\xfa\xd5\xe6\xb0\x23\x96\x1c\x28\x53\xd6\x95\x45\x79\x8b\xd6\x11\x9b\x6e\xf5\xb4\x05\xc4\x85\x3d\x69\xdf\x12\xdf\x1e\xb0\x3f\xc1\x61\x58\xb5\xa7\xb2\xca\x29\xfd\x4f\x3c\x76\x20\x70\x4e\xb6\x9d\x05\xc4\x80\x2f\xf8\x5f\xd6\xe5\x76\x81\x6c\x47\x52\x86\xa3\x81\x49\x76\x81\x6c\x47\xb2\x1a\x3a\x06\xe6\xd9\x11\xb5\xa5\x3d\x65\xec\x18\xda\xa0\x5d\x30\x6b\x32\x7b\x55\x98\x42\x51\x88\xe0\x90\xaa\x9d\x57\x5f\x1a\x89\x00\x18\x16\x8a\x62\x12\xc1\x1f\x61\xc3\xa1\x49\x51\x16\xe1\x7e\xac\x62\x4f\x5b\xbc\x1f\x83\x54\xbf\xfb\x31\x99\xc3\x42\x54\x5b\x12\xbe\xb5\xde\x06\xaf\x94\xb9\xf6\x50\x89\x7a\xb8\xa0\x11\xa0\x89\x80\xe5\x18\x9d\x0b\x08\x9f\x72\xb4\xbb\x9a\xbe\xaf\x0b\xa5\xee\x90\x27\x49\xee\xf7\xd1\xc6\x3d\x14\x05\x5c\xc5\xac\x51\x36\xcb\x6f\x51\xd4\x75\xdf\xf5\xfb\x6b\xdb\x9c\x98\x8d\xc1\x38\x78\xb9\x6c\x96\xa6\xc0\x16\x3c\x6d\x50\x03\xe7\x3e\x82\xab\x10\xd6\xe2\xfd\xdc\x41\x3b\x1a\x97\xe3\x76\x9e\x9d\xa7\xb0\xc2\x58\xe5\x0e\xe1\x36\x43\xf3\x31\x21\xf7\x78\x1c\x02\xc8\xc1\x86\x9c\x23\xf3\xd0\x63\x9a\x6e\x8b\x7c\x01\x65\xe9\x7e\x54\x46\x25\xa4\x44\x88\x7c\x14\xc5\x44\xe0\xa9\xbc\x8f\x17\x3d\x08\xf4\xd5\xd4\x70\x24\xac\xbe\xc4\xdb\x68\x3d\xad\xa5\x6b\x43\x88\x13\x8c\x1f\xa5\xd5\xad\xba\x6e\xf9\xdf\xa2\x72\x6c\x8e\xd0\xf5\xc3\xdb\x60\xf0\x97\x91\x3f\x9a\x6c\xc5\x3e\x01\x5e\xbf\xd8\x29\x89\x4e\xd6\x64\xc6\x95\x1c\x58\xc5\xc1\x00\x27\xbb\x48\xf7\xc9\x63\x9f\x8d\x5c\xb8\x5a\xa4\xa8\x1c\x1e\x47\x74\xf8\x38\xbf\xf9\xf0\xeb\xdf\xd7\x8b\x9b\xb9\xd4\x35\x32\x10\x3d\xb9\xc7\xcc\x72\xe6\x20\x37\x1a\x6d\x50\xee\x76\xce\xe3\x06\xe6\xb7\xbf\xcd\x40\x93\xc5\xd8\xb3\xdd\x45\x75\x39\xe1\xbb\x52\x18\xc4\x08\x4f\xb2\x37\x52\x19\x85\x4b\x12\x65\x96\x3d\x4f\xbf\x5e\x35\x6a\x53\xba\xe3\xeb\xdd\x6c\xbd\xb0\x2e\x97\x85\xb7\xb8\x0f\x9f\x93\xbf\x5d\xcc\xde\x97\xa7\xfc\x46\x26\xfc\x1f\x0f\x50\xfb\xd4\xf4\x6c\xb6\xab\x1e\x87\xea\xf9\x5f\xdf\x2d\xe7\xc3\x04\xdf\xf7\xd0\x3c\xc8\x91\x3d\x3a\x2f\x17\x39\xbc\x8f\x29\x17\x14\x66\xea\x98\x04\xd4\x4b\xbc\x35\xa5\x38\x74\xa4\xbd\x7a\xda\x62\xec\xd1\x5f\xa3\x7f\x03\x00\x00\xff\xff\x46\xb6\x0f\x2c\x3e\x1b\x00\x00")
+var _wski18nResourcesEn_usAllJson = []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\xcc\x98\x51\x6f\xdb\x36\x10\xc7\xdf\xfd\x29\x0e\x7a\x71\x06\xb8\xfa\x00\xdd\x53\xb0\x19\x73\xb0\xae\x31\x56\x67\x7d\x58\x86\x81\x16\xcf\xd6\x21\x12\xc9\x92\x94\x3d\xd7\xd0\x77\x1f\x48\xd9\xb5\xd7\x48\x96\x44\x2b\x59\x9e\x62\x30\xbc\xff\xfd\x78\x3c\x1e\x8f\xfa\x73\x04\xb0\x1f\x01\x00\x44\xc4\xa3\xf7\x10\x3d\x08\xb6\xcc\x10\xac\x04\xc6\x39\x68\x59\x58\x04\xa9\x2c\x49\x61\x60\xbc\xdf\xc7\x87\xdf\x65\x39\x8e\x26\x95\x9d\xd5\x4c\x98\x8c\xb9\xe1\x16\x81\xf7\x70\x2e\x10\x8d\x00\xca\x49\xb3\xff\x44\x23\xb3\x08\xb3\xc5\x62\x0e\x1a\xbf\x14\x68\x2c\xac\xa4\x86\xf9\xc3\xc2\x93\x78\xe9\xb2\x1c\x7b\x55\xd4\xba\x2c\x5b\x89\x02\x24\x03\x21\x7f\x99\x0e\x0e\x79\x41\x32\x10\xf2\xe7\xe9\x87\xe9\x62\x3a\x34\xe7\x65\xd5\xd0\x4d\xbf\xff\x34\xfc\xae\x5f\xd0\x6c\xc1\x64\x4a\xa1\xe0\x0d\x07\xc3\x4d\x78\xf8\xfd\xc3\x21\xf7\x03\xa1\xaf\xf7\xd0\x2d\xd2\xc7\x80\x38\x39\x27\x54\xe8\x2c\x28\xba\xad\x3a\xb5\x38\x77\x62\xc3\x32\xe2\xa1\x14\x9d\xcd\x6b\x9d\x4f\xb5\x96\x1a\x50\x24\x92\x93\x58\x7f\x13\x59\x4a\xbe\x6b\xf5\xdc\xcd\xf6\x82\x5b\x12\x64\x89\x65\xf4\xf5\xcc\xbc\xa3\xd7\x16\xd3\xb6\xd4\xe5\x1c\x6c\x7a\x38\x0e\xac\xb0\x29\x0a\x4b\x89\x77\x01\x29\x32\x8e\xba\x4f\x96\xf6\x11\xab\x05\xbb\x2d\x6c\x2a\x35\x7d\xad\x6c\x9e\x70\x07\x64\x40\x48\x0b\x89\x14\x2b\x5a\x17\x1a\x39\xdc\xbc\x7b\xe7\xb4\xdd\x7f\xdc\x72\x49\x23\xff\xa1\x01\x2d\x58\xae\x1e\x4e\xc0\xed\xfc\x0e\x52\x69\x2c\xe4\x85\xdb\x5f\x04\xa5\xe5\x86\x38\xf2\xf8\x51\x34\x31\xb4\x58\x75\xd8\xa0\xd7\xbf\x77\x7f\x92\x79\xce\x04\x87\x15\xa3\x0c\x39\xf0\xa2\x92\x12\xd5\xd6\xba\xd1\x42\x63\x83\xeb\x6e\xb6\xb5\x6e\x3f\x4a\x20\x61\x51\xaf\x58\x72\x0a\xd2\x8f\x20\xe4\xb1\x5c\x1b\x25\x85\x41\x7f\xb0\x00\xff\x51\x98\x58\xe4\x0d\x18\x61\x5a\xfd\xa2\xe1\x1d\x08\x96\x85\x46\xe4\x99\x7d\xad\xfb\x45\x8a\xb0\x92\x59\x26\xb7\xee\x88\x33\xa5\xb2\xe3\xa1\x42\x5f\x01\xb6\xcc\xa5\x6e\x82\xb4\x41\xde\x7a\x5a\x03\xc5\xde\x5e\xbd\x7e\xb3\x35\xec\xa4\xe5\x16\xa4\x98\x36\xd5\xa5\xbc\x41\x6d\x48\x8a\x7e\xf7\x69\x07\x89\x2b\x7b\xd2\xd0\x2b\xbe\xbb\x60\x38\xe0\x30\x54\xdd\x51\x96\x05\x65\xff\xc9\xc7\x1e\x00\x97\x6c\xbb\x45\xc0\x05\xf0\x19\xff\x75\x5d\x6e\x1f\xc9\x6e\x90\xee\x71\x34\x30\x64\x1f\xc9\x6e\x90\x87\x47\xc7\xc0\x9c\x3d\x55\x3b\xc6\xd3\x3d\x3b\x86\x0e\x68\x1f\xcd\x86\xca\x7e\xb8\x98\xfc\xa5\x10\xc3\xb1\x54\x1b\xcb\xbe\x35\x12\x5e\xd0\x0f\x94\xe5\x38\x86\x3f\xfc\x84\x63\x93\xc2\x34\xc2\x63\xc4\x12\x4b\x1b\x7c\x8c\xc0\xdd\x7e\x8f\x11\x89\xe3\x40\xdc\x78\x25\xbc\xb4\xdf\x96\x5d\xa9\x6a\xed\xf1\x26\x0a\xd8\x82\x56\x81\x36\x00\x2d\x13\x34\xc6\x2b\x7c\x29\x50\xef\x1a\xfa\xbe\x3e\x48\xfd\x25\x6b\x21\xf7\xfb\x38\x37\xeb\xb2\x84\x9b\x44\x72\x74\x93\xdd\xdf\xb2\x6c\xea\xbe\x9b\xe7\x37\xb6\x39\x89\x14\x02\x13\xbf\xcb\x55\xb3\x34\x01\xa9\xc1\x52\x8e\x1c\x64\x61\x63\xb8\xf1\x69\xed\x76\xbf\x30\xd0\x0d\xe3\x7a\xdd\xde\x6f\xe7\x09\x2c\x31\x61\x85\x41\xb8\x57\x28\x3e\xa7\x64\x9e\x4e\x8f\x00\x32\x90\x93\x31\x24\xd6\x01\xaf\xe9\xae\xca\x57\x20\xbb\xee\x87\x29\xaa\x24\x5d\x86\xb8\x1f\x65\x39\x76\xf2\x54\x9d\xc7\xab\x3e\x08\x84\x7a\xaa\x5d\xd2\xc5\x28\xc0\xcd\x3c\x43\x66\xf0\xf4\xdc\x83\xcf\xb3\xbb\x4f\xbf\xfe\x7d\x3b\xbf\x9b\xb9\x1a\x49\x02\xe2\xad\x79\x52\x5a\x2a\x03\x85\xe0\xa8\x3d\x93\xd9\x19\x8b\x39\xcc\xee\x7f\x9b\x02\x27\x8d\x89\x95\x7a\x17\x37\xe5\xd7\xab\x22\x0c\x12\x84\xad\x9b\x1b\x33\x45\x3e\xe0\xb1\xd2\xd2\xca\xc9\xf7\xa3\x82\xe5\xe8\xeb\xe7\xf7\xb3\xa5\xb6\x8e\xba\x1a\x76\xdc\xa8\x2d\xe1\x39\xfc\xfd\x7c\xfa\xb1\x5a\xe5\x0b\x85\xf0\x7f\x5c\x40\xe3\x67\x8b\xb3\x77\xc2\xe1\x43\x43\x33\xff\xed\xc3\x62\x36\x4c\xf2\xbd\x86\xe7\x41\x96\x6c\xd1\x58\x12\xeb\xd8\x7f\x6b\x61\xc6\x3b\x54\xcc\xa6\x20\x57\x55\x1d\x78\xae\xb7\xa2\x0c\x87\xce\xb4\x37\x8f\xed\x82\x3d\xfa\x6b\xf4\x6f\x00\x00\x00\xff\xff\xc3\xb5\x3e\xea\x8a\x19\x00\x00")
 
 func wski18nResourcesEn_usAllJsonBytes() ([]byte, error) {
     return bindataRead(
@@ -129,7 +129,7 @@ func wski18nResourcesEn_usAllJson() (*asset, error) {
         return nil, err
     }
 
-    info := bindataFileInfo{name: "wski18n/resources/en_US.all.json", size: 6974, mode: os.FileMode(420), modTime: time.Unix(1506610567, 0)}
+    info := bindataFileInfo{name: "wski18n/resources/en_US.all.json", size: 6538, mode: os.FileMode(420), modTime: time.Unix(1506031828, 0)}
     a := &asset{bytes: bytes, info: info}
     return a, nil
 }
@@ -149,7 +149,7 @@ func wski18nResourcesEs_esAllJson() (*asset, error) {
         return nil, err
     }
 
-    info := bindataFileInfo{name: "wski18n/resources/es_ES.all.json", size: 0, mode: os.FileMode(420), modTime: time.Unix(1500156160, 0)}
+    info := bindataFileInfo{name: "wski18n/resources/es_ES.all.json", size: 0, mode: os.FileMode(420), modTime: time.Unix(1506031186, 0)}
     a := &asset{bytes: bytes, info: info}
     return a, nil
 }
@@ -169,7 +169,7 @@ func wski18nResourcesFr_frAllJson() (*asset, error) {
         return nil, err
     }
 
-    info := bindataFileInfo{name: "wski18n/resources/fr_FR.all.json", size: 0, mode: os.FileMode(420), modTime: time.Unix(1500156160, 0)}
+    info := bindataFileInfo{name: "wski18n/resources/fr_FR.all.json", size: 0, mode: os.FileMode(420), modTime: time.Unix(1506031186, 0)}
     a := &asset{bytes: bytes, info: info}
     return a, nil
 }
@@ -189,7 +189,7 @@ func wski18nResourcesIt_itAllJson() (*asset, error) {
         return nil, err
     }
 
-    info := bindataFileInfo{name: "wski18n/resources/it_IT.all.json", size: 0, mode: os.FileMode(420), modTime: time.Unix(1500156160, 0)}
+    info := bindataFileInfo{name: "wski18n/resources/it_IT.all.json", size: 0, mode: os.FileMode(420), modTime: time.Unix(1506031186, 0)}
     a := &asset{bytes: bytes, info: info}
     return a, nil
 }
@@ -209,7 +209,7 @@ func wski18nResourcesJa_jaAllJson() (*asset, error) {
         return nil, err
     }
 
-    info := bindataFileInfo{name: "wski18n/resources/ja_JA.all.json", size: 0, mode: os.FileMode(420), modTime: time.Unix(1500156160, 0)}
+    info := bindataFileInfo{name: "wski18n/resources/ja_JA.all.json", size: 0, mode: os.FileMode(420), modTime: time.Unix(1506031186, 0)}
     a := &asset{bytes: bytes, info: info}
     return a, nil
 }
@@ -229,7 +229,7 @@ func wski18nResourcesKo_krAllJson() (*asset, error) {
         return nil, err
     }
 
-    info := bindataFileInfo{name: "wski18n/resources/ko_KR.all.json", size: 0, mode: os.FileMode(420), modTime: time.Unix(1500156160, 0)}
+    info := bindataFileInfo{name: "wski18n/resources/ko_KR.all.json", size: 0, mode: os.FileMode(420), modTime: time.Unix(1506031186, 0)}
     a := &asset{bytes: bytes, info: info}
     return a, nil
 }
@@ -249,7 +249,7 @@ func wski18nResourcesPt_brAllJson() (*asset, error) {
         return nil, err
     }
 
-    info := bindataFileInfo{name: "wski18n/resources/pt_BR.all.json", size: 0, mode: os.FileMode(420), modTime: time.Unix(1500156160, 0)}
+    info := bindataFileInfo{name: "wski18n/resources/pt_BR.all.json", size: 0, mode: os.FileMode(420), modTime: time.Unix(1506031186, 0)}
     a := &asset{bytes: bytes, info: info}
     return a, nil
 }
@@ -269,7 +269,7 @@ func wski18nResourcesZh_hansAllJson() (*asset, error) {
         return nil, err
     }
 
-    info := bindataFileInfo{name: "wski18n/resources/zh_Hans.all.json", size: 0, mode: os.FileMode(420), modTime: time.Unix(1500156160, 0)}
+    info := bindataFileInfo{name: "wski18n/resources/zh_Hans.all.json", size: 0, mode: os.FileMode(420), modTime: time.Unix(1506031186, 0)}
     a := &asset{bytes: bytes, info: info}
     return a, nil
 }
@@ -289,7 +289,7 @@ func wski18nResourcesZh_hantAllJson() (*asset, error) {
         return nil, err
     }
 
-    info := bindataFileInfo{name: "wski18n/resources/zh_Hant.all.json", size: 0, mode: os.FileMode(420), modTime: time.Unix(1500156160, 0)}
+    info := bindataFileInfo{name: "wski18n/resources/zh_Hant.all.json", size: 0, mode: os.FileMode(420), modTime: time.Unix(1506031186, 0)}
     a := &asset{bytes: bytes, info: info}
     return a, nil
 }
